+++ aws-runtime/aws-config/common/src/aws/sdk/kotlin/runtime/config/profile/AwsSharedConfig.kt
@@ -5,7 +5,9 @@
 
 package aws.sdk.kotlin.runtime.config.profile
 
+import aws.sdk.kotlin.runtime.ConfigurationException
 import aws.sdk.kotlin.runtime.InternalSdkApi
+import aws.smithy.kotlin.runtime.net.Url
 
 /**
  * Represents shared configuration (profiles, SSO sessions, credentials, etc)
@@ -29,8 +31,29 @@ public class AwsSharedConfig internal constructor(
         get() = sections[ConfigSectionType.SSO_SESSION] ?: emptyMap()
 
     /**
+     * Map of section name to ServiceConfig
+     */
+    public val services: Map<String, ServicesConfig>
+        get() = sections[ConfigSectionType.SERVICES] ?: emptyMap()
+
+    /**
      * Resolve the active profile or the default profile if none is defined
      */
     public val activeProfile: AwsProfile
         get() = profiles[source.profile] ?: AwsProfile(source.profile, emptyMap())
 }
+
+/**
+ * Get the configured endpoint URL for a specific service, falling back to the global default.
+ * @param serviceKey The config key for the service, generally this is sdkId in snake_case form.
+ */
+public fun AwsSharedConfig.resolveEndpointUrl(serviceKey: String): Url? =
+    resolveServiceEndpointUrl(serviceKey) ?: activeProfile.endpointUrl
+
+private fun AwsSharedConfig.resolveServiceEndpointUrl(serviceKey: String): Url? =
+    activeProfile.servicesSection?.let { sectionName ->
+        val section = services[sectionName]
+            ?: throw ConfigurationException("shared config points to nonexistent services section '$sectionName'")
+
+        section.getUrlOrNull(serviceKey, "endpoint_url")
+    }